refactor(vue-wrapper): remove vue-wrapper and its testing
EX-1885
