--- packages/react-wrapper/src/__tests__/stubs/react-in-vue/react-destroy.stub.tsx
@@ -1,17 +1,0 @@
-import React from 'react';
-
-interface ReactDestroyedProps {
-  count: number
-}
-
-export const reactDestroyedCallback = jest.fn();
-
-export class ReactDestroyed extends React.Component<ReactDestroyedProps> {
-  componentWillUnmount(): void {
-    reactDestroyedCallback();
-  }
-
-  render() {
-    return <p>{ this.props.count }</p>;
-  }
-}
