--- packages/react-wrapper/src/__tests__/stubs/react-in-vue/react-paragraph-message.stub.tsx
@@ -1,13 +1,0 @@
-import React from 'react';
-
-export interface MessageProps {
-  message: string;
-}
-
-export function ParagraphMessage(props: MessageProps) {
-  return (
-    <p>
-      { props.message ?? 'Hello world!' }
-    </p>
-  );
-}
