--- packages/react-wrapper/src/__tests__/stubs/react-in-vue/react-title-message.stub.tsx
@@ -1,13 +1,0 @@
-import React from 'react';
-
-export interface MessageProps {
-  message: string;
-}
-
-export function TitleMessage(props: MessageProps) {
-  return (
-    <h1>
-      { props.message ?? 'Hello world!' }
-    </h1>
-  );
-}
