--- packages/react-wrapper/src/__tests__/stubs/react-in-vue/vue-title-message.stub.ts
@@ -1,13 +1,0 @@
-import Vue, { ComponentOptions } from 'vue';
-import { VueWrapper } from '../../../vue-wrapper';
-import { TitleMessage } from './react-title-message.stub';
-
-export const VueTitleMessage: ComponentOptions<Vue> = {
-  render(h) {
-    return h(VueWrapper, {
-      props: {
-        reactComponent: TitleMessage
-      }
-    });
-  }
-}
