+++ packages/react-wrapper/src/__tests__/stubs/react-simple-message.stub.tsx
@@ -1,5 +1,5 @@
 import React from 'react';
-import { ReactWrapper } from '../../../react-wrapper';
+import { ReactWrapper } from '../../react-wrapper';
 import { VueMessage } from './vue-message.stub';
 
 export function ReactSimpleMessageStub() {
