+++ packages/react-wrapper/src/__tests__/stubs/vue-destroy.stub.ts
