+++ packages/react-wrapper/src/__tests__/stubs/vue-message.stub.ts
