+++ packages/react-wrapper/src/__tests__/stubs/vue-scoped-slot.stub.ts
