+++ packages/react-wrapper/src/__tests__/stubs/vue-slots.stub.ts
